chore: reorganize cli package for future extension  (#2611)
* chore: git mv packages/cli -> packages/catalyst

* chore: move test files

* chore: move files into new entrypoint folder

* chore: change linting rules around to account for moved files
