+++ packages/catalyst/.eslintrc.cjs
@@ -4,12 +4,18 @@
 const config = {
   root: true,
   extends: ['@bigcommerce/catalyst/base', '@bigcommerce/catalyst/prettier'],
+  ignorePatterns: ['/dist/**'],
   rules: {
-    'no-console': 'off',
-    'import/no-named-as-default': 'off',
     '@typescript-eslint/naming-convention': 'off',
   },
-  ignorePatterns: ['/dist/**', '/templates/**'],
+  overrides: [
+    {
+      files: ['./src/cli/**'],
+      rules: {
+        'no-console': 'off',
+      },
+    },
+  ],
 };
 
 module.exports = config;
