+++ packages/catalyst/src/cli/commands/deploy.spec.ts
@@ -15,6 +15,12 @@ import {
   vi,
 } from 'vitest';
 
+import { server } from '../../../tests/mocks/node';
+import { textHistory } from '../../../tests/mocks/spinner';
+import { consola } from '../lib/logger';
+import { mkTempDir } from '../lib/mk-temp-dir';
+import { program } from '../program';
+
 import {
   createDeployment,
   deploy,
@@ -23,15 +29,10 @@ import {
   getDeploymentStatus,
   parseEnvironmentVariables,
   uploadBundleZip,
-} from '../../src/commands/deploy';
-import { consola } from '../../src/lib/logger';
-import { mkTempDir } from '../../src/lib/mk-temp-dir';
-import { program } from '../../src/program';
-import { server } from '../mocks/node';
-import { textHistory } from '../mocks/spinner';
+} from './deploy';
 
 // eslint-disable-next-line import/dynamic-import-chunkname
-vi.mock('yocto-spinner', () => import('../mocks/spinner'));
+vi.mock('yocto-spinner', () => import('../../../tests/mocks/spinner'));
 
 let exitMock: MockInstance;
