+++ packages/catalyst/src/cli/commands/start.spec.ts
@@ -2,9 +2,10 @@ import { Command } from 'commander';
 import { execa } from 'execa';
 import { afterEach, beforeAll, beforeEach, expect, test, vi } from 'vitest';
 
-import { start } from '../../src/commands/start';
-import { consola } from '../../src/lib/logger';
-import { program } from '../../src/program';
+import { consola } from '../lib/logger';
+import { program } from '../program';
+
+import { start } from './start';
 
 vi.mock('node:fs', () => ({
   existsSync: vi.fn(() => true),
