+++ packages/catalyst/src/cli/lib/get-module-cli-path.ts
