+++ packages/catalyst/src/cli/lib/project-config.ts
