+++ packages/catalyst/src/cli/lib/wrangler-config.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from 'vitest';
 
-import { getWranglerConfig } from '../../src/lib/wrangler-config';
+import { getWranglerConfig } from './wrangler-config';
 
 test('returns a config with name identical to worker self reference service', () => {
   const config = getWranglerConfig('uuid', 'kv-namespace-id');
