+++ packages/catalyst/templates/open-next.config.ts
@@ -1,10 +1,9 @@
-import { defineCloudflareConfig } from '@opennextjs/cloudflare';
+import { defineCloudflareConfig, type OpenNextConfig } from '@opennextjs/cloudflare';
 import { purgeCache } from '@opennextjs/cloudflare/overrides/cache-purge/index';
 import kvIncrementalCache from '@opennextjs/cloudflare/overrides/incremental-cache/kv-incremental-cache';
 import doQueue from '@opennextjs/cloudflare/overrides/queue/do-queue';
 import queueCache from '@opennextjs/cloudflare/overrides/queue/queue-cache';
 import doShardedTagCache from '@opennextjs/cloudflare/overrides/tag-cache/do-sharded-tag-cache';
-import { type OpenNextConfig } from '@opennextjs/cloudflare';
 
 const cloudflareConfig = defineCloudflareConfig({
   incrementalCache: kvIncrementalCache,
