+++ packages/catalyst/tsup.config.ts
@@ -1,7 +1,9 @@
 import { defineConfig, Options } from 'tsup';
 
 export default defineConfig((options: Options) => ({
-  entry: ['src/index.ts'],
+  entry: {
+    cli: 'src/cli/index.ts',
+  },
   format: ['esm'],
   clean: !options.watch,
   sourcemap: true,
